Move test to old file.
